+++ src/orangecontrib/ewoksdemo/sumtask.py
@@ -1,7 +1,7 @@
 from ewokscore.tests.examples.tasks.sumtask import SumTask
 
-from ewoksorange.gui.orange_utils.orange_imports import Input
-from ewoksorange.gui.orange_utils.orange_imports import Output
+from ewoksorange.gui.orange_utils.signals import Input
+from ewoksorange.gui.orange_utils.signals import Output
 from ewoksorange.gui.owwidgets.threaded import OWEwoksWidgetOneThread
 from ewoksorange.gui.widgets.simple_types_mixin import IntegerAdderMixin
 
@@ -14,16 +14,9 @@ class OWSumTask(IntegerAdderMixin, OWEwoksWidgetOneThread, ewokstaskclass=SumTas
     icon = "icons/sum.png"
     want_main_area = True
 
-    if Input is None:
-        inputs = [("A", object, ""), ("B", object, "")]
-        outputs = [{"name": "A + B", "id": "A + B", "type": object}]
-        inputs_orange_to_ewoks = {"A": "a", "B": "b"}
-        outputs_orange_to_ewoks = {"A + B": "result"}
-    else:
+    class Inputs:
+        a = Input("A", object)
+        b = Input("B", object)
 
-        class Inputs:
-            a = Input("A", object)
-            b = Input("B", object)
-
-        class Outputs:
-            result = Output("A + B", object)
+    class Outputs:
+        result = Output("A + B", object)